+++ webviz_subsurface/plugins/__init__.py
@@ -61,6 +61,7 @@ from ._surface_with_grid_cross_section import SurfaceWithGridCrossSection
 from ._surface_with_seismic_cross_section import SurfaceWithSeismicCrossSection
 from ._swatinit_qc import SwatinitQC
 from ._tornado_plotter_fmu import TornadoPlotterFMU
+from ._vfp_analysis import VfpAnalysis
 from ._volumetric_analysis import VolumetricAnalysis
 from ._well_analysis import WellAnalysis
 from ._well_completion import WellCompletion