+++ webviz_subsurface/plugins/_vfp_analysis/_types.py
@@ -1,0 +1,14 @@
+from webviz_config.utils import StrEnum
+
+
+class PressureType(StrEnum):
+    BHP = "BHP"
+    DP = "DP"
+
+
+class VfpParam(StrEnum):
+    THP = "THP"
+    WFR = "WFR"
+    GFR = "GFR"
+    ALQ = "ALQ"
+    RATE = "RATE"