+++ webviz_subsurface/plugins/_vfp_analysis/_utils/__init__.py
@@ -1,0 +1,1 @@
+from ._vfp_data_model import VfpDataModel, VfpTable