+++ webviz_subsurface/plugins/_vfp_analysis/_views/_vfp_view/_settings/__init__.py
@@ -1,0 +1,4 @@
+from ._filters import Filters
+from ._pressure_option import PressureOption
+from ._selections import Selections
+from ._vizualisation import Vizualisation