+++ webviz_subsurface/plugins/_vfp_analysis/_views/_vfp_view/_utils/__init__.py
@@ -1,0 +1,1 @@
+from ._vfp_figure_builder import VfpFigureBuilder