+++ webviz_subsurface/plugins/_vfp_analysis/_views/_vfp_view/_view_elements/__init__.py
@@ -1,0 +1,1 @@
+from ._vfp_graph import VfpGraph